-Moves EventCompression to separate class from event repo
