+++ app/src/main/java/com/andrew/tasky/agenda/data/event/EventRepositoryImpl.kt
@@ -4,7 +4,6 @@ import android.content.Context
 import com.andrew.tasky.R
 import com.andrew.tasky.agenda.data.database.AgendaDatabase
 import com.andrew.tasky.agenda.data.storage.ImageStorage
-import com.andrew.tasky.agenda.data.util.BitmapConverters
 import com.andrew.tasky.agenda.data.util.ModifiedType
 import com.andrew.tasky.agenda.domain.EventRepository
 import com.andrew.tasky.agenda.domain.models.AgendaItem
@@ -15,10 +14,6 @@ import com.andrew.tasky.auth.util.getResourceResult
 import com.andrew.tasky.core.Resource
 import com.andrew.tasky.core.UiText
 import javax.inject.Inject
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.async
-import kotlinx.coroutines.supervisorScope
-import kotlinx.coroutines.withContext
 import kotlinx.serialization.encodeToString
 import kotlinx.serialization.json.Json
 import okhttp3.MultipartBody
@@ -33,11 +28,11 @@ class EventRepositoryImpl @Inject constructor(
 ) : EventRepository {
 
     private val imageStorage = ImageStorage(context)
-    private val bitmapConverter = BitmapConverters
+    private val eventCompressor = EventCompressor(context)
 
     override suspend fun upsertEvent(event: AgendaItem.Event): UpsertEventResult {
 
-        val (compressedEvent, photosDeleted) = compressEvent(event)
+        val (compressedEvent, photosDeleted) = eventCompressor.compressEvent(event)
 
         val isEventInDb = db.getEventDao().getEventById(compressedEvent.id) == null
         val isModifiedCreateEvent = db.getEventDao().getModifiedEventById(compressedEvent.id)
@@ -199,47 +194,6 @@ class EventRepositoryImpl @Inject constructor(
         }
     }
 
-    private data class CompressEventResult(val event: AgendaItem.Event, val photosDeleted: Int)
-    private suspend fun compressEvent(event: AgendaItem.Event): CompressEventResult {
-        return withContext(Dispatchers.Default) {
-            var photosDeleted = 0
-            val compressedEvent = event.copy(
-                photos = supervisorScope {
-                    event.photos
-                        .chunked(MAX_PARALLEL_IMAGE_COMPRESS_COUNT)
-                        .map { eventPhotoChunk ->
-                            eventPhotoChunk.map { eventPhoto ->
-                                async {
-                                    if (eventPhoto is LocalPhoto) {
-                                        if (eventPhoto.savedInternally) {
-                                            val byteArray =
-                                                imageStorage.getByteArray(eventPhoto.key)
-                                            eventPhoto.copy(byteArray = byteArray)
-                                        } else {
-                                            val byteArray = eventPhoto.bitmap?.let { bitmap ->
-                                                bitmapConverter.bitmapToCompressByteArray(
-                                                    bitmap, MAX_PHOTO_SIZE_IN_BYTES
-                                                )
-                                            }
-                                            if (byteArray != null) {
-                                                eventPhoto.copy(byteArray = byteArray)
-                                            } else {
-                                                photosDeleted++
-                                                null
-                                            }
-                                        }
-                                    } else {
-                                        eventPhoto
-                                    }
-                                }
-                            }.mapNotNull { it.await() }
-                        }.flatten()
-                }
-            )
-            CompressEventResult(event = compressedEvent, photosDeleted = photosDeleted)
-        }
-    }
-
     private suspend fun saveLocalPhotoInternally(photo: LocalPhoto) {
         photo.byteArray?.let {
             ImageStorage(context = context).saveImage(
@@ -247,10 +201,5 @@ class EventRepositoryImpl @Inject constructor(
                 byteArray = it
             )
         }
-    }
-
-    companion object {
-        const val MAX_PHOTO_SIZE_IN_BYTES = 1_000_000
-        private const val MAX_PARALLEL_IMAGE_COMPRESS_COUNT = 3
     }
 }
